+++ app/src/main/java/com/jeiel85/wildhavenidle/presentation/home/HomeViewModel.kt
@@ -74,4 +74,10 @@ class HomeViewModel(
             gameRepository.markOnboardingCompleted()
         }
     }
+
+    fun claimDailyBonus() {
+        viewModelScope.launch {
+            gameRepository.claimDailyBonus()
+        }
+    }
 }